requirements.txt
asgiref==3.8.1
asttokens==3.0.0
colorama==0.4.6
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
executing==2.2.0
icecream==2.1.4
pillow==11.1.0
psycopg2==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
